refactor(auth): update AuthSession type location
