+++ src/modules/auth/domain/auth-session.types.ts
