+++ src/modules/auth/domain/auth.respository.ts
@@ -1,5 +1,5 @@
 import type { RepositoryResult } from '@/shared/domain/repository.result';
-import type { AuthSession } from '@/modules/auth/domain/errors/auth-session.types';
+import type { AuthSession } from '@/modules/auth/domain/auth-session.types';
 
 export interface IAuthRepository {
   getSession(): RepositoryResult<AuthSession | null>;
