+++ src/modules/auth/infrastructure/auth.mapper.ts
@@ -1,5 +1,5 @@
 import type { Session as NextAuthSession } from 'next-auth';
-import type { AuthSession } from '@/modules/auth/domain/errors/auth-session.types';
+import type { AuthSession } from '@/modules/auth/domain/auth-session.types';
 
 export class AuthMapper {
   static toDomain(data: NextAuthSession): AuthSession {
