+++ src/modules/auth/mocks/auth.mocks.repository.ts
@@ -1,7 +1,7 @@
 import { Success } from '@/shared/domain/result';
 import type { RepositoryResult } from '@/shared/domain/repository.result';
 import type { IAuthRepository } from '@/modules/auth/domain/auth.respository';
-import type { AuthSession } from '@/modules/auth/domain/errors/auth-session.types';
+import type { AuthSession } from '@/modules/auth/domain/auth-session.types';
 
 /**
  * @property session: An `AuthSession` | `null` type to control fake-active session
